ENH: remove tracking.tools in favor of utils
